fix models import
